requirements.txt
openai==1.109.1
python-docx==1.2.0
pyfiglet==1.0.4

